Stamina regeneration

+++ CharacterStats.cs
@@ -11,7 +11,7 @@ namespace SOULS
         public int currentHealth;
 
         public int staminaLevel = 10;
-        public int maxStamina;
-        public int currentStamina;
+        public float maxStamina;
+        public float currentStamina;
     }
 }
